requirements.txt
langchain==1.2.6
python-dotenv==1.2.1
fastmcp==2.14.3
langchain-mcp-adapters==0.2.1

